Added QC checks to annotate tool. Restructured code format to match analyze.
