+++ reditools/annotate/__init__.py
@@ -1,0 +1,3 @@
+from .main import main
+
+__all__ = ('main',)
